refactor(utils): move URITypeAdapter out of databind to break databind cycle
URITypeAdapter is a generic Gson TypeAdapter<URI> with no project dependencies.
Living in databind, it was the sole reason codelenses/commands/inlayhints data
classes imported the databind package, while databind imports those features to
register their JSON subtypes — a package cycle.

Relocating the adapter to the leaf utils package removes every inbound edge to
databind (it now only depends outward, on the features it registers), so
databind leaves the package cycle (import-level SCC shrinks 22 -> 21).
