+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/utils/URITypeAdapter.java
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with BSL Language Server.
  */
-package com.github._1c_syntax.bsl.languageserver.databind;
+package com.github._1c_syntax.bsl.languageserver.utils;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
